RLE: code review fixes (rename bits to grids)
